Opo iki lali

+++ database/migrations/2023_12_09_102751_create_carts_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->unsignedBigInteger('user_id')->nullable();
             $table->foreign('user_id')->references('id')->on('users');
             $table->double('price');
-            $table->double('discount_item')->default(0);
+            $table->integer('discount_item')->default(0);
             $table->double('total');
             $table->integer('quantity')->nullable();
             $table->integer('jumlah_jual')->nullable();
